consolidate Logger and Logging modules
